### 2023-06-06 #5521 @editorialbot
[PRE REVIEW]: Tethys: a Spatiotemporal Downscaling Model for Global Water Demand
**Submitting author:** @ifthompson (<a href="http://orcid.org/0000-0001-9594-0043">Isaac Thompson</a>)
**Repository:** https://github.com/JGCRI/tethys
**Branch with paper.md** (empty if default branch): dev
**Version:** v2.0.0
