+++ paper/paper.md
@@ -7,7 +7,7 @@ tags:
 authors:
   - name: Isaac Thompson
     orcid: 0000-0001-9594-0043
-	corresponding: true
+    corresponding: true
     affiliation: 1
   - name: Zarrar Khan
     orcid: 0000-0002-8147-8553
