### 2023-06-06 #5521 @kthyng
Hi @ifthompson and thanks for your submission! Because the size of the package is relatively small (lines of code < 1000) I need to ping the editorial board to vote on whether the submission is in scope in terms of [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This will take 1-2 weeks, thanks for your patience.
