### 2023-06-08 8d14b04
generate faster gcamreader queries
