+++ tethys/datareader/regional.py
@@ -15,10 +15,12 @@ def load_region_data(gcam_db, sectors, demand_type='withdrawals'):
     dbpath, dbfile = os.path.split(gcam_db)
     conn = gcamreader.LocalDBConn(dbpath, dbfile)
 
+    search_sectors = [unfriendly_sector_name(i) for i in sectors]
+
     inputs = {'withdrawals': ['water_td_*_W', '*_water withdrawals', '*_desalinated water'],
               'consumption': ['water_td_*_C', '*_water consumption']}.get(demand_type)
 
-    df = conn.runQuery(easy_query('demand-physical', technology='!water_td_*', input=inputs))
+    df = conn.runQuery(easy_query('demand-physical', sector=search_sectors, technology='!water_td_*', input=inputs))
 
     # add '_BasinName' to region if exists
     df['region'] += df['sector'].apply(extract_basin_name) + df['input'].apply(extract_basin_name)
@@ -27,7 +29,7 @@ def load_region_data(gcam_db, sectors, demand_type='withdrawals'):
 
     # handle technology as sector
     df = df[df.sector.isin(sectors) | df.technology.isin(sectors)]
-    df.loc[~df.sector.isin(sectors), 'sector'] = df.technology[~df.sector.isin(sectors)]
+    df.loc[df.technology.isin(sectors), 'sector'] = df.technology[df.technology.isin(sectors)]
 
     df = df.groupby(['region', 'sector', 'year'])[['value']].sum().reset_index()
 
