### 2023-06-14 58da112
handle GCAM sector/subsector/technology better
