### 2023-07-12 95eefc6
helper functions for converting dataframes of points to grids
