### 2023-07-25 a611ae7
prevent invalid netcdf variable names
