### 2023-07-31 48b2ad3
make temporal downscaling methods more explicit about their inputs, but breaks earlier usage
