+++ tethys/tdmethods/irrigation.py
@@ -1,16 +1,13 @@
 from tethys.datareader.gridded import load_file
+from tethys.datareader.maps import load_region_map
 
 
-def temporal_distribution(model):
+def temporal_distribution(years, resolution, regionfile, irrfile, irrvar='pirrww'):
     """Temporal downscaling of irrigation water demand"""
 
-    years = range(model.years[0], model.years[-1] + 1)
-    irr = load_file(model.temporal_files['irr'], model.resolution, years, regrid_method='label')['pirrww']
+    irr = load_file(irrfile, resolution, years, regrid_method='label', variables=[irrvar])[irrvar]
 
-    irr_regions = model.inputs.region[(model.inputs.sector == 'Irrigation') &
-                                      (model.inputs.region.isin(model.region_masks.region.data))].unique()
-
-    region_masks = model.region_masks.sel(region=irr_regions)
+    region_masks = load_region_map(regionfile, masks=True, target_resolution=resolution)
 
     irr_grouped = irr.where(region_masks, 0)
     month_sums = irr_grouped.sum(dim=('lat', 'lon'))
