### 2023-07-31 da2073b
make temporal downscaling methods more explicit about their inputs, but breaks earlier usage
