### 2023-07-31 1dc65b0
add a generic temporal downscaling function to use some weight/proxy
