### 2023-07-31 50eb3fd
config option for spatial bounds
