### 2023-07-31 6440f39
config option for spatial bounds
