+++ tethys/datareader/maps.py
@@ -2,10 +2,10 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 
-from tethys.datareader.gridded import regrid
+from tethys.datareader.gridded import regrid, pad_global, crop
 
 
-def load_region_map(mapfile, masks=False, namefile=None, target_resolution=None, nodata=None, flip_lat=False):
+def load_region_map(mapfile, masks=False, namefile=None, target_resolution=None, bounds=None, nodata=None, flip_lat=False):
     """ Load region map.
 
     :param mapfile: path to map file
@@ -41,8 +41,12 @@ def load_region_map(mapfile, masks=False, namefile=None, target_resolution=None,
         da = da.squeeze('band').drop_vars('band')
 
     if target_resolution is not None:
+        da = pad_global(da)
         da = regrid(da, target_resolution, method='label')
         da = da.chunk(chunks=dict(lat=-1, lon=-1))
+
+    if bounds is not None:
+        da = crop(da, bounds)
 
     # use region names provided in CSV namefile, otherwise check metadata for a names dict
     if namefile is not None:
