### 2023-08-01 91648d9
new output writing
