### 2023-08-01 bf186db
new output writing
