### 2023-08-01 8281977
add spatial bounds option to temporal downscaling methods
