+++ tethys/tdmethods/irrigation.py
@@ -2,12 +2,12 @@ from tethys.datareader.gridded import load_file
 from tethys.datareader.maps import load_region_map
 
 
-def temporal_distribution(years, resolution, regionfile, irrfile, irrvar='pirrww'):
+def temporal_distribution(years, resolution, regionfile, irrfile, irrvar='pirrww', bounds=None):
     """Temporal downscaling of irrigation water demand"""
 
-    irr = load_file(irrfile, resolution, years, regrid_method='label', variables=[irrvar])[irrvar]
+    irr = load_file(irrfile, resolution, years, bounds=bounds, regrid_method='label', variables=[irrvar])[irrvar]
 
-    region_masks = load_region_map(regionfile, masks=True, target_resolution=resolution)
+    region_masks = load_region_map(regionfile, masks=True, target_resolution=resolution, bounds=bounds)
 
     irr_grouped = irr.where(region_masks, 0)
     month_sums = irr_grouped.sum(dim=('lat', 'lon'))
