### 2023-08-01 08e12b3
add spatial bounds option to temporal downscaling methods
