### 2023-08-02 b1d0592
if single layer file but multiple variables declared, use that for all
