### 2023-08-02 f6d15bf
if single layer file but multiple variables declared, use that for all
