### 2023-08-03 3fae35a
add supersector_iterations config option to control how many times to repeat applying total and individual sector region-scale constraints
