### 2023-08-04 #5521 @ifthompson
Hi @jsta 👋 , the [conceptual overview](https://jgcri.github.io/tethys/user_guide.html#conceptual-overview) section of the docs explains some of the ideas in the paper in more detail. Usually it is most convenient to control operation with a [configuration file](https://jgcri.github.io/tethys/user_guide.html#configuration-file), and then just run `tethys.run_model('configuration_file.yml)`. I'm happy to discuss the code more if you have specific questions. Also, pinging coauthor @zarrarkhan here who can also answer questions.

I don't know a lot of people so I don't have any suggested reviewers in mind. Can you pick volunteers?
