### 2023-08-09 #5521 @jsta
:wave: @eldemet Would you be willing to review this submission to JOSS?
