### 2023-08-09 #5521 @jsta
:wave: @mariosmsk Would you be willing to review this submission to JOSS?
