### 2023-08-23 #5521 @nickrsan
@jsta thanks for thinking of me. I'd be willing to and the software looks interesting, but I'll note that downscaling isn't my area of expertise. I can possibly recommend another reviewer who knows Python and downscaling if you'd prefer.
