### 2023-08-23 #5521 @nickrsan
The person I know who would have expertise in these areas is @khegewisch, if she's available.  I'd still be happy to review the parts of it that are in my wheelhouse if you need a reviewer, but if Katherine has time/interest, I think she could better evaluate the methods, in addition to expertise in the software components.
