### 2023-09-15 #5521 @nickrsan
Yes, I'm still willing, but likely can't review until at least October due to current commitments. If that's acceptable, I'd be happy to review the paper.
