### 2023-09-18 #5855 @editorialbot
[REVIEW]: Tethys: a Spatiotemporal Downscaling Model for Global Water Demand
**Submitting author:** @ifthompson (<a href="http://orcid.org/0000-0001-9594-0043">Isaac Thompson</a>)
**Repository:** https://github.com/JGCRI/tethys
**Branch with paper.md** (empty if default branch): main
**Version:** v2.1.0
