### 2023-10-31 #5855 @Mariosmsk
Hi @ifthompson,

Some comments so far.

> License: Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved]
(https://opensource.org/licenses/alphabetical) software license?

I recommend updating the License file to include an OSI-approved software license. Additionally, please ensure that the `setup.py` file correctly references the chosen license. This will help ensure compliance with open-source licensing standards.

>  Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

I checked the usage example, I have the image after spatial downscaling but not the image before. It's essential to also provide the original image before downscaling.

>  Installation: Does installation proceed as outlined in the documentation?

Installation proceeded as outlined in the documentation. However, I recommend adding `matplotlib` to the requirements.txt file to ensure all dependencies are clearly documented and can be easily installed. Regarding the Dask example, I encountered Java-related errors, even though I believe I have added Java to the path. For better reproducibility, consider including code in the example that automatically configures the Java path. This will improve the user experience and minimize potential issues during execution.
