### 2023-12-19 #5855 @nickrsan
Just making a note for the license checkbox since GitHub isn't picking up the license from its format and advertising it by name - it appears to be a 2-clause BSD license - not anything the authors need to respond to, just to mark this for the review.
