### 2023-12-21 JGCRI/tethys#61 @nickrsan
JOSS Review: Missing item under documentation section for how to get support
Hi there,

I'm one of the reviewers for the JOSS paper on this package - openjournals/joss-reviews#5855. I've encountered one item that appears incomplete:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) **Seek support**

I don't see in the README or documentation any notes for where someone would seek support. If this already exists, can you point me to it, and otherwise, I'd consider this a minor modification that needs addressing before I can check that item off.

Thank you!
