### 2023-12-21 #5855 @nickrsan
Hi all,

As a progress update, I've filed a few issues - mostly minor, but blocking further review, on the tethys repository (see cross-references above from GitHub). Also, I'll note that two unit tests are currently failing on my machine (Win 11, Python 3.10) - though I think one is that I probably need to set the JAVA path (`test_regional_query`) and the other I'll dig into and file an issue if appropriate (`test_easy_query` not producing expected output).
