### 2024-01-11 #5855 @jsta
@crvernon Do you have any insight into Issac's availability that might inform this review?
