+++ .github/workflows/docs.yml
@@ -1,9 +1,13 @@
+---
 name: "docs"
-on:
+
+on:    # yamllint disable-line rule:truthy
   push:
-    branches:
-      - dev
-      - main
+    branches-ignore:
+      - 'false'
+  pull_request:
+    branches-ignore:
+      - 'false'
 
 jobs:
   docs:
