+++ setup.py
@@ -46,7 +46,7 @@ setup(
             'sphinx-panels~=0.6.0',
             'sphinx-rtd-theme~=0.5.2',
             'sphinx-mathjax-offline~=0.0.1',
-            'twine~=3.4.1'
+            'twine~=3.4.1',
             'click-default-group>=1.2.4'
         ]
     }
