+++ .github/workflows/build.yml
@@ -33,7 +33,8 @@ jobs:
         - name: Install dependencies
           run: |
             python -m pip install --upgrade pip
-            python setup.py install
+            pip install click-default-group
+            pip install .
 
         - name: Test and generate coverage report on Linux
           run: |
