### 2024-02-21 #5855 @crvernon
@kthyng yes, I am going to add myself as the listed author on this.  I will keep Isaac as a shared first author and will adjust the paper accordingly.

@nickrsan and @Mariosmsk thank you for your patience!  I will be taking over this review so please reach out if you have any questions or issues.  I have caught up on the issues you have raised and will begin to systematically address those.  I'll link all activity to this thread so we can have a single point of access to all of our conversations.
