### 2024-02-21 b90756d
fix error in easy query
