### 2024-02-21 385d986
updated process for downloading example data
