### 2024-02-21 f9e38aa
fix license specification
