+++ requirements.txt
@@ -7,3 +7,7 @@ dask>=2022.12.1
 xarray>=2022.09.0
 rioxarray>=0.12.4
 tqdm>=4.66.2
+matplotlib>=3.8.3
+dask[distributed]>=2024.2.0
+bokeh!=3.0.*,>=2.4.2
+
