### 2024-04-12 794f546
add hvplot dep
