+++ requirements.txt
@@ -3,11 +3,12 @@ gcamreader>=1.2.5
 numpy>=1.22
 pandas>=1.2.4
 netCDF4>=1.6
-dask>=2022.12.1
 xarray>=2022.09.0
 rioxarray>=0.12.4
 tqdm>=4.66.2
 matplotlib>=3.8.3
-dask[distributed]>=2024.2.0
+dask[complete]>=2022.12.1
 bokeh!=3.0.*,>=2.4.2
-
+geopandas>=0.14.3
+dask_geopandas>=0.3.1
+hvplot>=0.9.2
