+++ tethys/__init__.py
@@ -1,4 +1,6 @@
 from .model import *
 from .utils.install_supplement import get_example_data, default_download_dir
+from .utils.spatial import InputViewer
 
-__version__ = "2.0.1"
+
+__version__ = "2.1"
