+++ .github/workflows/paper.yml
@@ -1,5 +1,5 @@
 ---
-name: paper
+name: "paper"
 
 on:    # yamllint disable-line rule:truthy
   push:
