+++ paper/paper.md
@@ -9,7 +9,7 @@ authors:
     orcid: 0000-0001-9594-0043
     affiliation: 1
     equal-contrib: true
-  - name: Chris Vernon
+  - name: Chris R. Vernon
     orcid: 0000-0002-3406-6214
     affiliation: 2
     corresponding: true
@@ -28,19 +28,19 @@ bibliography: paper.bib
 
 # Summary
 
-Humans use water for many important tasks, such as drinking, growing food, and cooling power plants. Since future water demands depend on complex global interactions between economic sectors (e.g., demand for wheat in one country causing demand for water to grow that wheat in another country), it is often modeled at coarse spatial and temporal scales as part of models that account for complex, multi-sector systems dynamics. However, models that project future water availability typically simulate physical processes at much finer scales. `Tethys` enables coupling between these kinds of models by downscaling region-scale water demand projections using sector-specific proxies and formulas.
+Humans use water for many important tasks, such as drinking, growing food, and cooling power plants. Since future water demands depend on complex global interactions between economic sectors (e.g., demand for wheat in one country causing demand for water to grow that wheat in another country), it is often modeled at coarse spatial and temporal scales as part of models that account for complex, multi-sector system dynamics. However, models that project future water availability typically simulate physical processes at much finer scales. `Tethys` enables integration between these kinds of models by downscaling region-scale water demand projections using sector-specific proxies and formulas.
 
-# Statement of need
+# Statement of Need
 
-Hydrological models require finely gridded water demand data to represent the location and timing of flows into the human system, but historical inventories of water use are often only available per country, and at annual or larger intervals [@hess-22-2117-2018]. In order to model future global economic linkages in detail, multi-sector dynamics models also operate at these coarser spatial and temporal scales. This gap in scale makes downscaling water demands a common problem.
+Global hydrological models often require gridded water demand data to represent the location and timing of flows for human consumption, but historical inventories of water use are often only available per country at annual or larger intervals [@hess-22-2117-2018]. In order to model future global economic linkages in detail, multi-sector models (e.g., the Global Change Analysis Model (GCAM; [@gmd-12-677-2019;@gmd-15-2533-2022])) also operate at these coarser spatial and temporal scales. This gap in scale makes downscaling water demands a common need.
 
-The distribution of water demands depends on the location and timing of activities that use water, so the usual approach is to use relevant gridded datasets as spatial proxies for each water demand sector (e.g., assume that irrigation water demand is proportional to irrigated land area), then further allocate annual water demands among months according to formulas that capture seasonal variations [@hess-17-4555-2013]. This is typically accomplished with scripts designed for specific model-coupling workflows, but different models and proxy datasets can have different breakdowns of water demand sectors, limiting reuse of such scripts.
+The distribution of water demands depends on the location and timing of activities that use water, so the usual approach is to use relevant gridded datasets as spatial proxies for each water demand sector (e.g., assume that irrigation water demand is proportional to irrigated land area), then further allocate annual water demands among months according to formulas that capture seasonal variations [@hess-17-4555-2013]. This is typically accomplished with scripts designed for specific model-integration workflows, but different models and proxy datasets can have different breakdowns of water demand sectors, limiting reuse of such scripts.
 
[…]
