+++ paper/paper.bib
@@ -50,3 +50,26 @@ PAGES = {4555--4575},
 URL = {https://hess.copernicus.org/articles/17/4555/2013/},
 DOI = {10.5194/hess-17-4555-2013}
 }
+@Article{gmd-12-677-2019,
+AUTHOR = {Calvin, K. and Patel, P. and Clarke, L. and Asrar, G. and Bond-Lamberty, B. and Cui, R. Y. and Di Vittorio, A. and Dorheim, K. and Edmonds, J. and Hartin, C. and Hejazi, M. and Horowitz, R. and Iyer, G. and Kyle, P. and Kim, S. and Link, R. and McJeon, H. and Smith, S. J. and Snyder, A. and Waldhoff, S. and Wise, M.},
+TITLE = {GCAM v5.1: representing the linkages between energy, water, land, climate,
+and economic systems},
+JOURNAL = {Geoscientific Model Development},
+VOLUME = {12},
+YEAR = {2019},
+NUMBER = {2},
+PAGES = {677--698},
+URL = {https://gmd.copernicus.org/articles/12/677/2019/},
+DOI = {10.5194/gmd-12-677-2019}
+}
+@Article{gmd-15-2533-2022,
+AUTHOR = {Binsted, M. and Iyer, G. and Patel, P. and Graham, N. T. and Ou, Y. and Khan, Z. and Kholod, N. and Narayan, K. and Hejazi, M. and Kim, S. and Calvin, K. and Wise, M.},
+TITLE = {GCAM-USA v5.3\_water\_dispatch: integrated modeling of subnational US energy, water, and land systems within a global framework},
+JOURNAL = {Geoscientific Model Development},
+VOLUME = {15},
+YEAR = {2022},
+NUMBER = {6},
+PAGES = {2533--2559},
+URL = {https://gmd.copernicus.org/articles/15/2533/2022/},
+DOI = {10.5194/gmd-15-2533-2022}
+}
