+++ paper/paper.md
@@ -32,7 +32,7 @@ Humans use water for many important tasks, such as drinking, growing food, and c
 
 # Statement of Need
 
-Global hydrological models often require gridded water demand data to represent the location and timing of flows for human consumption, but historical inventories of water use are often only available per country at annual or larger intervals [@hess-22-2117-2018]. In order to model future global economic linkages in detail, multi-sector models (e.g., the Global Change Analysis Model (GCAM; [@gmd-12-677-2019;@gmd-15-2533-2022])) also operate at these coarser spatial and temporal scales. This gap in scale makes downscaling water demands a common need.
+Global hydrological models often require gridded water demand data to represent the location and timing of flows for human consumption, but historical inventories of water use are often only available per country at annual or larger intervals [@hess-22-2117-2018]. In order to model future global economic linkages in detail, multi-sector models (e.g., the Global Change Analysis Model [@gmd-12-677-2019;@gmd-15-2533-2022]) also operate at these coarser spatial and temporal scales. This gap in scale makes downscaling water demands a common need.
 
 The distribution of water demands depends on the location and timing of activities that use water, so the usual approach is to use relevant gridded datasets as spatial proxies for each water demand sector (e.g., assume that irrigation water demand is proportional to irrigated land area), then further allocate annual water demands among months according to formulas that capture seasonal variations [@hess-17-4555-2013]. This is typically accomplished with scripts designed for specific model-integration workflows, but different models and proxy datasets can have different breakdowns of water demand sectors, limiting reuse of such scripts.
 
@@ -40,7 +40,7 @@ Building on previous versions [@Li-2018], `Tethys` now generalizes this downscal
 
 # Key Functionality
 
-`Tethys` consists of 2 stages: spatial downscaling ((\autoref{fig:1})) and (optionally) temporal downscaling. First, sectoral water demands by region are disaggregated to water demand by grid cell in proportion to appropriate spatial proxies, i.e.,
+`Tethys` consists of 2 stages: spatial downscaling (\autoref{fig:1}) and (optionally) temporal downscaling. First, sectoral water demands by region are disaggregated to water demand by grid cell in proportion to appropriate spatial proxies, i.e.,
 
 \begin{equation}
 \text{demand}_\text{cell} = \text{demand}_\text{region} \times \frac{\text{proxy}_\text{cell}}{\text{proxy}_\text{region}}.
@@ -48,7 +48,7 @@ Building on previous versions [@Li-2018], `Tethys` now generalizes this downscal
 
 Then, temporal downscaling follows sector-specific formulas from the literature, which determine the fraction of a year’s water demand to allocate to each month based on relationships between monthly water demand and other monthly variables. See the [documentation](https://jgcri.github.io/tethys) for more details and example usage.
 
-![Before and after spatial downscaling \label{fig:1}](example.png)
+![Before and after spatial downscaling. \label{fig:1}](example.png)
 
 # Acknowledgements
 
