+++ paper/paper.bib
@@ -1,75 +1,78 @@
 @Article{hess-22-2117-2018,
-AUTHOR = {Huang, Z. and Hejazi, M. and Li, X. and Tang, Q. and Vernon, C. and Leng, G. and Liu, Y. and D\"oll, P. and Eisner, S. and Gerten, D. and Hanasaki, N. and Wada, Y.},
-TITLE = {Reconstruction of global gridded monthly sectoral water withdrawals for 1971--2010 and analysis of their spatiotemporal patterns},
-JOURNAL = {Hydrology and Earth System Sciences},
-VOLUME = {22},
-YEAR = {2018},
-NUMBER = {4},
-PAGES = {2117--2133},
-URL = {https://hess.copernicus.org/articles/22/2117/2018/},
-DOI = {10.5194/hess-22-2117-2018}
+    AUTHOR = {Huang, Z. and Hejazi, M. and Li, X. and Tang, Q. and Vernon, C. and Leng, G. and Liu, Y. and D\"oll, P. and Eisner, S. and Gerten, D. and Hanasaki, N. and Wada, Y.},
+    TITLE = {Reconstruction of global gridded monthly sectoral water withdrawals for 1971--2010 and analysis of their spatiotemporal patterns},
+    JOURNAL = {Hydrology and Earth System Sciences},
+    VOLUME = {22},
+    YEAR = {2018},
+    NUMBER = {4},
+    PAGES = {2117--2133},
+    URL = {https://hess.copernicus.org/articles/22/2117/2018/},
+    DOI = {10.5194/hess-22-2117-2018}
 }
+
 @article{Li-2018,
- abstract = {Downscaling of water withdrawals from regional/national to local scale is a fundamental step and also a common problem when integrating large scale economic and integrated assessment models with high-resolution detailed sectoral models. Tethys, an open-access software written in Python, is developed with statistical downscaling algorithms, to spatially and temporally downscale water withdrawal data to a finer scale. The spatial resolution will be downscaled from region/basin scale to grid (0.5 geographic degree) scale and the temporal resolution will be downscaled from year to month. Tethys is used to produce monthly global gridded water withdrawal products based on estimates from the Global Change Assessment Model (GCAM). Funding statement: PNNL is operated for DOE by Battelle Memorial Institute under contract DE-AC05-76RL01830},
- author = {Li, Xinya and Vernon, Chris R. and Hejazi, Mohamad I. and Link, Robert P. and Huang, Zhongwei and Liu, Lu and Feng, Leyang},
- doi = {10.5334/jors.197},
- journal = {Journal of Open Research Software},
- keyword = {en_US},
- month = {Feb},
- title = {Tethys – A Python Package for Spatial and Temporal Downscaling of Global Water Withdrawals},
- year = {2018}
+    author = {Li, Xinya and Vernon, Chris R. and Hejazi, Mohamad I. and Link, Robert P. and Huang, Zhongwei and Liu, Lu and Feng, Leyang},
+    doi = {10.5334/jors.197},
+    journal = {Journal of Open Research Software},
+    keyword = {en_US},
+    month = {Feb},
+    title = {Tethys – A Python Package for Spatial and Temporal Downscaling of Global Water Withdrawals},
+    year = {2018}
 }
+
 @Article{Khan2023,
-author={Khan, Zarrar
-and Thompson, Isaac
-and Vernon, Chris R.
-and Graham, Neal T.
-and Wild, Thomas B.
-and Chen, Min},
-title={Global monthly sectoral water use for 2010--2100 at 0.5{\textdegree} resolution across alternative futures},
[…]
