### 2024-04-12 1370e7e
note on java for dashboard
