+++ .github/workflows/docs.yml
@@ -1,36 +1,0 @@
----
-name: "docs"
-
-on:    # yamllint disable-line rule:truthy
-  push:
-    branches:
-      - 'main'
-  pull_request:
-    branches:
-      - 'main'
-
-jobs:
-  docs:
-    runs-on: ubuntu-latest
-    steps:
-    - name: Install latex
-      run: |
-        sudo apt-get update -y && sudo apt-get install -y texlive-latex-extra dvipng
-    - uses: actions/checkout@v2
-    - uses: actions/setup-python@v2
-      with:
-        python-version: '3.9'
-    - name: Update pip and install python dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install -e .[dev]
-    - name: Build html docs
-      working-directory: 'docs/'
-      run: |
-        make html
-    - name: Deploy
-      uses: JamesIves/github-pages-deploy-action@4.1.4
-      with:
-        branch: gh-pages # The branch the action should deploy to.
-        folder: docs/build/html # The folder the action should deploy. 
-
