### 2024-04-12 ec0e8eb
remove docs badge
