### 2024-04-12 #5855 @crvernon
:wave: @Mariosmsk

Addressing your comments...all fixes are now available on `main`.  I have release a new version (2.1) with the following changes accounted for as well.  Thank you!

1) --------------------------------------------------

> > License: Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved]
> > (https://opensource.org/licenses/alphabetical) software license?
>
> I recommend updating the License file to include an OSI-approved software license. Additionally, please ensure that the `setup.py` file correctly references the chosen license. This will help ensure compliance with open-source licensing standards.

No problem. I reformatted the BSD 2 Simplified license to be recognized by GitHub and also added in the correct SPDX short identifier in setup.py.  This was done in https://github.com/JGCRI/tethys/pull/65

2) --------------------------------------------------

> > Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.
>
> I checked the usage example, I have the image after spatial downscaling but not the image before. It's essential to also provide the original image before downscaling.

The data that gets downscaled is from the input CSV file.  You can see this in the example data and it gets referenced in the configuration file that is used to run **tethys**.  This tabular data gets allocated spatially to this raster in the example (also in the example data):  `data/maps/regions.tif`.

Once you have downloaded the example data via `tethys.get_example_data()` you can run the following to see what is actually in the example configuration file (or you can open it up in an editor):

```python
import tethys

# assuming you downloaded to the default location
config_file = tethys.default_download_dir + '/example/config_example.yml'

with open(config_file) as cfg:
    print(cfg.read())
```
this prints...
```
# Basic example config file for Tethys

# Project level settings
years: [2010, 2015, 2020, 2025, 2030]
resolution: 0.125
demand_type: withdrawals

# output settings
output_file: output/example_output.nc

# csv input
csv: data/example_data.csv

# dictionary of proxy files and promised variables and years
proxy_files:
  data/population/ssp2_{year}.tif:
    variables: Population
    years: [2010, 2020, 2030]

# mapping of inputs sectors to proxies
downscaling_rules:
  Municipal: Population

# list of map files
map_files:
  - data/maps/regions.tif
```

You can see the input CSV file which holds the data to be downscaled in `csv` and the raster containing the regions as the first item in `map_files`.

All of this information is available when you run **tethys** via the variable you store your run initialization in.  For example:

```python
# run tethys using the config file
result = tethys.run_model(config_file)
```
You can use `result` to get anything in the configuration, like the `csv` path:

```python
result.csv
```
returns:
```
'data/example_data.csv'
```

Based on your comment, I decided to build plotting functionality within the package that allows you to produce an interactive visualization the input data to be downscaled with the regional spatial data.  So you can now do the following using the `result` object from your Tethys run:

```python
from tethys import InputViewer

# instantiate input viewer
gridded_inputs = InputViewer(
    result,
    sector="Municipal",
    year=2020,
)

# plot interactive figure
gridded_inputs.plot(cnorm="log")
```
Of course, you can also call help on this method to see the options available:
```python
help(gridded_inputs.plot)
```
I also added in a method to allow the dataset to be exported to a raster file, so:
```python
gridded_inputs.to_raster("<your raster path>")
```

This was done in https://github.com/JGCRI/tethys/pull/68

3) --------------------------------------------------

> > Installation: Does installation proceed as outlined in the documentation?
>
> Installation proceeded as outlined in the documentation. However, I recommend adding `matplotlib` to the requirements.txt file to ensure all dependencies are clearly documented and can be easily installed. Regarding the Dask example, I encountered Java-related errors, even though I believe I have added Java to the path. For better reproducibility, consider including code in the example that automatically configures the Java path. This will improve the user experience and minimize potential issues during execution.

I completely agree concerning the package dependencies.  This was fixed in https://github.com/JGCRI/tethys/pull/68
For the Java path configuration, I chose to place a notice to the user above that example section.  These types of path configurations can vary based on OS and package manager utilization and can sometimes cause more grief to the user.  But thank you so much for highlighting this as a need!  This was done in https://github.com/JGCRI/tethys/pull/69
