### 2024-04-12 #5855 @crvernon
:wave: @nickrsan - some of your comments were addressed in the above response, but the others that were specifically addressed where done in the following:

- https://github.com/JGCRI/tethys/issues/62 was addressed in https://github.com/JGCRI/tethys/pull/64
- https://github.com/JGCRI/tethys/issues/61 was addressed in https://github.com/JGCRI/tethys/pull/63

Thanks so much for these great comments!
