+++ CITATION.cff
@@ -4,7 +4,7 @@ authors:
 - family-names: "Thompson"
   given-names: "Isaac"
   orcid: "https://orcid.org/0000-0001-9594-0043"
-- family-names: "Vernonv
+- family-names: "Vernon"
   given-names: "Chris R."
   orcid: "https://orcid.org/0000-0002-3406-6214"
 - family-names: "Khan"
