+++ CITATION.cff
@@ -10,4 +10,4 @@ authors:
 - family-names: "Khan"
   given-names: "Zarrar"
   orcid: "https://orcid.org/0000-0002-8147-8553"
-title: "Tethys: A Spatiotemporal Downscaling Model for Global Water Demand"
+title: "tethys: a spatiotemporal downscaling model for global water demand"
