+++ CITATION.cff
@@ -1,13 +1,0 @@
-cff-version: 1.2.0
-message: "If you use this software in your work, please cite it as below."
-authors:
-- family-names: "Thompson"
-  given-names: "Isaac"
-  orcid: "https://orcid.org/0000-0001-9594-0043"
-- family-names: "Vernon"
-  given-names: "Chris R."
-  orcid: "https://orcid.org/0000-0002-3406-6214"
-- family-names: "Khan"
-  given-names: "Zarrar"
-  orcid: "https://orcid.org/0000-0002-8147-8553"
-title: "tethys: a spatiotemporal downscaling model for global water demand"
