### 2024-05-17 #5855 @nickrsan
OK, this looks great and like a nicely polished package. I've checked off my checklist and think it's ready to go.

I'll note that I had one issue, which is that I couldn't get the Dask example to run, but could run the main example. Tracking it down a bit more (by trying to run the example_demeter.yml file without the dask config, as well as running the test suite), I get the impression that maybe there's an issue in the `gcamreader` dependency - java is set up on this machine and available on the path - Java is throwing some errors when being executed though. The failure being somewhere in gcamreader is suggested to me by the test_regional/test_load_region_data test failing on my machine too. I'd consider this to not affect the core claims of the package though, but wanted to flag this for you. I could provide more information in an issue on the tethys repo if you think it's something you want to track down.
