### 2024-05-18 #5855 @crvernon
Thanks @nickrsan!  I'll note your issue in the `gcamreader` repository and have someone check that out.  Thanks!

Also, thanks for your review and feedback. Much appreciated!

@kthyng I think we are good to go now.
